Fix Linux build: FFmpeg API compatibility and Qt type conversion
Problem 1 - FFmpeg channel layout API:
- Ubuntu 22.04 uses FFmpeg 4.x (older API)
- Uses codecpar->channels instead of ch_layout.nb_channels
- macOS Homebrew uses FFmpeg 7.x (newer API)

Solution:
- Add version check: LIBAVCODEC_VERSION_MAJOR >= 59
- Use ch_layout.nb_channels for FFmpeg 5.0+
- Use channels for FFmpeg 4.x

Problem 2 - Qt JSON type ambiguity:
- int64_t to QJsonValue conversion ambiguous in Qt 6.5.3
- Multiple constructors match: bool, int, qint64, double

Solution:
- Explicitly cast to qint64: static_cast<qint64>(getBitrate())
- Removes ambiguity for Qt's type system

Tested on:
- macOS with FFmpeg 7.1.1 (newer API)
- Should work on Linux with FFmpeg 4.x (older API)

+++ src/cli/main_cli.cpp
@@ -396,7 +396,7 @@ QJsonObject CLIProcessor::videoInfoToJson(VideoDecoder* decoder) {
     root["frame_rate"] = decoder->getFrameRate();
     root["duration"] = decoder->getDuration();
     root["total_frames"] = decoder->getFrameCount();
-    root["bitrate"] = decoder->getBitrate();
+    root["bitrate"] = static_cast<qint64>(decoder->getBitrate());
     root["pixel_format"] = decoder->getPixelFormat();
     return root;
 }
